Relocate authorization code from QuestionController to QuestionService when update or delete question.
question을 수정 또는 삭제할 때 현재 사용자가 작성자인지 QuestionService 내부에서 확인하도록 변경

+++ src/main/java/com/codessquad/qna/service/QuestionService.java
@@ -22,15 +22,23 @@ public class QuestionService {
         questionRepository.save(new Question(user, title, contents));
     }
 
-    public void updateQuestion(long questionId, String title, String contents) {
+    public void updateQuestion(long questionId, String title, String contents, User tryToUpdate) {
         Question question = getQuestion(questionId);
-        question.update(title, contents);
-        questionRepository.save(question);
+        if (question.getWriter().getId() == tryToUpdate.getId()) {
+            question.update(title, contents);
+            questionRepository.save(question);
+            return;
+        }
+        throw new IllegalStateException("자신의 글만 수정할 수 있습니다.");
     }
 
-    public void deleteQuestion(long questionId) {
+    public void deleteQuestion(long questionId, User tryToDelete) {
         Question question = getQuestion(questionId);
-        questionRepository.delete(question);
+        if (question.getWriter().getId() == tryToDelete.getId()) {
+            questionRepository.delete(question);
+            return;
+        }
+        throw new IllegalStateException(("자신의 글만 삭제할 수 있습니다."));
     }
 
     public List<Question> getQuestions() {
